Read a FASTQ formatted file

Read a FASTQ formatted file

Usage

read_fastq(fastq, path = NULL, format = "illumina1.8+")

Arguments

fastq

string, containing the FASTQ file name. Accepts compressed .gz files as well.

path

string, optional path to FASTQ file.

format

string, containing the FASTQ format. Options include: illumina1.8+, solexa, sanger. Default = illumina1.8+.

Value

A data.frame containing the FASTQ sequence data

Examples

fastq_data <- strollur::read_fastq(strollur_example("tiny.fastq.gz"))

# fastq_data is a data.frame.
# To access the names of the sequences in the file, run the following:

fastq_data$sequence_name
#> [1] "M00967:43:000000000-A3JHG:1:1101:18327:1699"
#> [2] "M00967:43:000000000-A3JHG:1:1101:14069:1827"
#> [3] "M00967:43:000000000-A3JHG:1:1101:18044:1900"

# To access the sequences in the file, run the following:

fastq_data$sequence
#> [1] "NACGGAGGATGCGAGCGTTATCCGGATTTATTGGGTTTAAAGGGTGCGTAGGCGGCCTGCCAAGTCAGCGGTAAAATTGCGGGGCTCAACCCCGTACAGCCGTTGAAACTGCCGGGCTCGAGTGGGCGAGAAGTATGCGGAATGCGTGGTGTAGCGGTGAAATGCATAGATATCACGCAGAACCCCGATTGCGAAGGCAGCATACCGGCGCCCTACTGACGCTGAGGCACGAAAGTGCGGGGATCAAACAG"
#> [2] "TACGGAGGATGCGAGCGTTATCCGGATTTATTGGGTTTAAAGGGTGCGTAGGCGGCCTGCCAAGTCAGCGGTAAAATTGCGGGGCTCAACCCCGTACAGCCGTTGAAACTGCCGGGCTCGAGTGGGCGAGAAGTATGCGGAATGCGTGGTGTAGCGGTGAAATGCATAGATATCACGCAGAACCCCGATTGCGAAGGCAGCATACCGGCGCCCTACTGACGCTGAGGCACGAAAGTGCGGGGATCAAACAG"
#> [3] "TACGGAGGATGCGAGCGTTGTCCGGAATCACTGGGCGTAAAGGGCGCGTAGGCGGTTTAATAAGTCAGTGGTGAAAACTGAGGGCTCAACCCTCAGCCTGCCACTGATACTGTTAGACTTGAGTATGGAAGAGGAGAATGGAATTCCTAGTGTAGCGGTGAAATGCGTAGATATTAGGAGGAACACCAGTGGCGAAGGCGATTCTCTGGGCCAAGACTGACGCTGAGGCGCGAAAGCGTGGGGAGCAAACA"

# To access the quality scores in the file, run the following:

fastq_data$quality_score
#> [[1]]
#>   [1]  2 29 29 32 32 33 32 33 33 37 37 37 38 38 38 38 38 38 38 38 38 38 38 38 38
#>  [26] 38 38 38 39 39 39 39 39 39 39 38 38 38 39 39 39 39 39 38 39 38 38 38 38 38
#>  [51] 38 38 39 38 38 38 38 38 38 39 39 39 39 39 39 39 39 39 39 38 38 38 38 38 39
#>  [76] 39 39 39 38 39 38 38 38 38 38 38 39 39 33 38 39 38 35 38 38 38 38 38 39 39
#> [101] 39 38 38 38 38 39 39 39 39 39 39 39 39 38 38 38 38 38 39 38 31 35 39 39 38
#> [126] 39 36 38 38 38 38 38 38 33 37 38 38 36 38 38 38 38 38 38 38 38 13 35 37 36
#> [151] 37 37 37 37 37 37 37 35 34 37 37 37 37 37 37 37 37 37 37 37 37 37 37 37 37
#> [176] 37 37 37 37 37 37 37 37 37 37 35 37 35 37 37 37 36 37 37 34 37 37 30 37 35
#> [201] 37 37 37 37 37 37 37 37 32 37 37 37 37 37 37 37 37 37 37 37 33 35 35 37 37
#> [226] 37 37 37 36 37 32 35 37 37 37 37 37 33 32 37 37 37 32 30 36 37 37 37 33 37
#> [251] 37
#> 
#> [[2]]
#>   [1] 18 32 32 30 32 33 33 35 33 37 37 33 36 38 38 36 38 38 38 38 32 37 37 38 38
#>  [26] 38 38 38 39 39 39 34 38 38 38 38 38 38 39 37 38 39 38 38 34 37 35 36 37 38
#>  [51] 38 38 39 38 38 38 36 38 37 16 38 38 37 38 39 39 39 39 39 38 38 36 38 38 39
#>  [76] 39 39 39 39 37 38 38 38 38 38 38 39 39 39 39 39 38 38 38 38 34 35 35 38 39
#> [101] 39 38 38 38 37 39 39 39 39 39 39 39 39 34 35 31 34 34 39 38 38 38 38 39 36
#> [126] 39 38 38 38 31 38 37 38 38 38 38 38 38 38 31 33 38 38 38 36 38 34 36 33 37
#> [151] 37 37 33 37 37 33 26 24 31 36 37 37 37 36 37 37 37 37 37 37 37 37 37 37 37
#> [176] 37 32 37 33 33 33 37 37 37 37 37 33 33 33 37 37 37 37 33 33 33 37 37 37 37
#> [201] 37 37 37 37 37 37 37 33 33 33 33 33 33 33 37 37 37 37 37 37 37 37 37 35 35
#> [226] 37 32 37 37 37 37 37 13 32 37 24 14 37 33 33 33 33 33 13 36 32 37 37 36 30
#> [251] 37
#> 
#> [[3]]
#>   [1] 33 32 31 33 33 33 32 33 33 37 37 37 38 38 38 38 38 38 38 38 38 38 39 38 38
#>  [26] 38 38 38 39 39 39 39 38 39 39 39 38 38 38 38 39 39 39 38 38 38 38 38 38 38
#>  [51] 38 38 39 38 38 38 38 38 38 37 39 37 39 39 39 37 39 39 39 39 39 39 38 37 39
#>  [76] 39 39 39 39 39 37 39 38 38 37 38 39 39 38 38 39 38 38 39 39 37 39 39 39 39
#> [101] 39 39 39 39 39 39 39 39 39 39 38 39 39 39 38 39 35 39 39 39 39 39 39 39 39
#> [126] 39 38 39 35 38 37 37 39 39 39 35 39 39 39 39 39 39 39 39 39 39 39 39 37 39
#> [151] 39 39 35 38 39 39 38 38 38 38 38 39 39 39 39 39 38 38 38 38 36 38 38 38 38
#> [176] 38 33 37 15 33 38 38 38 38 38 38 38 38 37 37 37 38 38 38 35 38 37 37 37 37
#> [201] 33 34 37 33 37 37 33 37 37 33 36 36 37 37 36 37 37 37 37 37 37 37 37 33 35
#> [226] 32 24 36 36 37 37 37 29 35 37 37 37 37 35 35 37 37 37 32 35 35 37 37 37 37
#> [251] 13
#>